backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
supabase==1.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.24.1
python-dotenv==1.0.0
alembic==1.12.1
email-validator==2.1.0
bcrypt==4.1.0